+++ aioxmpp/presence/__init__.py
@@ -23,13 +23,26 @@
 :mod:`~aioxmpp.presence` --- Peer presence bookkeeping
 ######################################################
 
-This module provides a :class:`Service` to track the presence of peers, no
-matter whether they are in the roster or not.
+This module provides a :class:`.PresenceClient` service to track the presence
+of peers, no matter whether they are in the roster or not.
 
 .. versionadded:: 0.4
 
-.. autoclass:: Service
+.. currentmodule:: aioxmpp
+
+.. autoclass:: PresenceService
+
+.. currentmodule:: aioxmpp.presence
+
+.. class:: Service
+
+   Alias of :class:`.PresenceClient`.
+
+   .. deprecated:: 0.8
+
+      The alias will be removed in 1.0.
 
 """
 
-from .service import Service  # NOQA
+from .service import PresenceClient  # NOQA
+Service = PresenceClient  # NOQA