+++ tests/disco/test___init__.py
@@ -21,14 +21,20 @@
 ########################################################################
 import unittest
 
+import aioxmpp
 import aioxmpp.disco as disco
 import aioxmpp.disco.service as disco_service
 import aioxmpp.disco.xso as disco_xso
 
 
 class TestExports(unittest.TestCase):
-    def test_Service(self):
-        self.assertIs(disco.Service, disco_service.Service)
+    def test_DiscoClient(self):
+        self.assertIs(disco.DiscoClient, disco_service.DiscoClient)
+        self.assertIs(aioxmpp.DiscoClient, disco_service.DiscoClient)
+
+    def test_DiscoServer(self):
+        self.assertIs(disco.DiscoServer, disco_service.DiscoServer)
+        self.assertIs(aioxmpp.DiscoServer, disco_service.DiscoServer)
 
     def test_xso(self):
         self.assertIs(disco.xso, disco_xso)