+++ tests/presence/test___init__.py
@@ -21,10 +21,17 @@
 ########################################################################
 import unittest
 
+import aioxmpp
 import aioxmpp.presence as presence
 import aioxmpp.presence.service as presence_service
 
 
 class TestExports(unittest.TestCase):
     def test_Service(self):
-        self.assertIs(presence.Service, presence_service.Service)
+        self.assertIs(presence.Service, presence_service.PresenceClient)
+
+    def test_RosterClient(self):
+        self.assertIs(presence.PresenceClient,
+                      presence_service.PresenceClient)
+        self.assertIs(aioxmpp.PresenceClient,
+                      presence_service.PresenceClient)